+++ tests/dummy/app/stereo-connections/local-dummy-connection.js
@@ -1,5 +1,6 @@
 import BaseSound from 'ember-stereo/stereo-connections/base';
 export default class LocalDummySound extends BaseSound {
+  static key = 'LocalDummyConnection';
   static toString() {
     return 'Local Dummy Connection';
   }