+++ CHANGELOG.md
@@ -6,10 +6,13 @@
 - [CHANGE] audio events now fire an object as the first argument, with the `sound`, matching the `play` and `load` functions. All event listeners should change from `this.stereo.on('event-name', (sound) => { // handler })` to `this.stereo.on('event-name', ({sound}) => { // handler })`
 - [CHANGE] Default volume is now 100
 - [CHANGE] Audio errors no longer throw errors that need to be caught, instead they are surfaced on the hifi service, or through the `{{sound-is-errored}}` and `{{sound-error-details}}`
+- [CHANGE] Sounds are cached ignoring query strings
+
 
 - [CHORE] Renamed anything that was `hifi` to `stereo`
 - [CHORE] Upgrade to Ember 3.24. This addon is octane-only now
 - [CHORE] Complex promise/event waiting loops that were prone to error and hard to test have been replaced with ember concurrency tasks
+
 
 # FORKED from ember-hifi